+++ window_scene/test/unittest/scene_session_test3.cpp
@@ -1227,8 +1227,8 @@ HWTEST_F(SceneSessionTest3, GetAllAvoidAreas, Function | SmallTest | Level2)
  */
 HWTEST_F(SceneSessionTest3, NotifyPipWindowSizeChange, Function | SmallTest | Level2)
 {
-    uint32_t width = 10;
-    uint32_t height = 10;
+    double width = 10;
+    double height = 10;
     double scale = 10;
     SessionInfo info;
     info.abilityName_ = "NotifyPipWindowSizeChange";
